Move interface.rs into util.rs
Also store if_index on XdpLink
Removes XdpLink::TryFrom<FdLink>
