+++ aya/src/lib.rs
@@ -47,7 +47,6 @@ extern crate bitflags;
 
 mod bpf;
 mod generated;
-pub mod interface;
 pub mod maps;
 mod obj;
 pub mod pin;
